requirements.txt
pandas>=1.5.0
matplotlib>=3.7.0
faiss-cpu>=1.7.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0
apify-client>=1.5.0
sec-api==1.0.33
numpy>=1.24.0
sentence-transformers>=2.2.0